server/requirements.txt
aniso8601==10.0.1
blinker==1.9.0
click==8.2.1
dotenv==0.9.9
Flask==3.1.1
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
packaging==25.0
pillow==11.3.0
playwright==1.53.0
pyee==13.0.0
PyMySQL==1.1.1
pytesseract==0.3.13
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
